Update on "[monarch] ActorStatus::Failed(String) -> Failed(ActorErrorKind)"
Refactor to get a more structured error to ActorStatus. This deletes the unused structure in ActorErrorKind to make it cloneable/serializable so that we can put it in ActorStatus.

Now that we can represent an ActorStatus that is an SupervisionEvent of a child, we can start to chain/collapse the supervision chain as it bubbles up.

Differential Revision: [D86723978](https://our.internmc.facebook.com/intern/diff/D86723978/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D86723978/)!

[ghstack-poisoned]