[hyperactor] refactor channel::net client loop
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1823

First, we break down the client loop into several functions, breaking up its monolithic nature.

Second, instead of relying on custom rendering of loop states, we formalize this into a span, which efficiently pulls out the (structured) fields as required. This means we don't pay a formatting/string construction overhead on each iteration, and also make available the underlying fields in our structured logs.

This is just the first salvo in improving how the `net` module is factored.
ghstack-source-id: 322824631
@exported-using-ghexport

Differential Revision: [D86783166](https://our.internmc.facebook.com/intern/diff/D86783166/)