Use pytest-isolate to make tests more reliable (#1528)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1528

The pytest-isolate package can support both pytest-timeout and pytest-forked,
to run each test in a separate process.
This should help prevent cross-contamination (one test affects a second test), as well
as preventing tests with fatal signals from taking down the whole test run.
It requires pytest-timeout to *not* be installed, or else that will take precedence.

Since this slows down tests, but makes them easier to reproduce, we can turn on `-n auto`
to run them in parallel.

https://pypi.org/project/pytest-isolate/

Differential Revision: D84636534