Update base for Update on "[hyperactor] support generic behaviors; implement Mesh resource behavior"
`hyperactor::behavior!` is expanded to support generic type parameters, allowing for parameterized behaviors.

We use this new functionality to formally define what a mesh controller behavior is, by bundling related types into a trait, so that we can write `Controller<SomeMeshType>`.

Mesh controllers should follow this behavior.

The plan is then to expand this to a set of generic code, tools, and so on, that can be parameterized over different implementations of `Mesh`.

Differential Revision: [D86420507](https://our.internmc.facebook.com/intern/diff/D86420507/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D86420507/)!

[ghstack-poisoned]