[hyperactor] assert_behaves!
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1826

This adds a macro that asserts an actor's behavior; we use it to test the mesh behavior:

```
hyperactor::assert_behaves!(TestMeshController as Controller<TestMesh>);
```

It is implemented by statically validating that the actor type can bind to the defined behavior. This then provide a way to both document the intended behavior of an actor, and also to statically validate at the definition site that the behavior is implemented correctly.

(We already statically validate this at the *use* site, but then the errors are already "at a distance".)
ghstack-source-id: 323322439

Differential Revision: [D86786231](https://our.internmc.facebook.com/intern/diff/D86786231/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D86786231/)!