Add InstanceState (#1888)
Summary:

There is a bug in `Instance`'s current Drop impl:

https://www.internalfb.com/code/fbsource/[5b70b5b30638d5db5dfef1cbf929dea6df7d3f56]/fbcode/monarch/hyperactor/src/proc.rs?lines=1594-1596

since we clone a `Instance` object on the python side:

https://www.internalfb.com/code/fbsource/[5b70b5b30638d5db5dfef1cbf929dea6df7d3f56]/fbcode/monarch/hyperactor/src/proc.rs?lines=1566

It means the Instance's actor status will be changed to `Stopped` when a clone is dropped, although the actor is still alive.

This diff adds a `InstanceState` struct, so the instance's lifetime can tracked by `Arc`'s ref counter. Then we move the `Drop` impl to this  `InstanceState`, so it will only be called when the ref count goes to 0.

Reviewed By: dulinriley

Differential Revision: D87075455