Option to build oss faster until we remove these
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1880

`controller` and `process_allocator` really are no longer needed to develop oss stuff, but we have to remove important testing, etc. For now, I just to be able to compile locally.

This is a 3x build speed improvement because each of these binaries uses different flags for the underlying crates and causes recompiles.
ghstack-source-id: 323494573
@exported-using-ghexport

Differential Revision: [D86880713](https://our.internmc.facebook.com/intern/diff/D86880713/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D86880713/)!