Add controller actor for meshes
Summary:
Fixes https://github.com/meta-pytorch/monarch/issues/1848

We want to support an important ownership axiom of Actors, that every actor which
is alive has a single alive owner.
If that owner is stopped (either voluntarily or via a crash), we want the child actors to
stop as well.

To this end, whenever we spawn a normal actor on a ProcMesh, we also spawn a "controller"
actor on the owner. This controller's only job at the moment is to ensure during cleanup
it calls "stop" on the actor it owns.
This should mean as long as your actor stops on its own, all of its children will recursively
stop.

This mechanism doesn't handle an unclean exit on the owner, in which case we wouldn't
be able to run the cleanup hook. That will need to be fixed separately.
It is an important case for when the client might just exit for some reason, or get Ctrl-C'd.

Had to enhance the `hyperactor::export` macro to handle generics. "hyperactor::remote"
still doesn't, but luckily we don't even need to do a remote spawn of this actor.

Reviewed By: mariusae

Differential Revision: D86917010