remote rust_local_mesh
This diff removes the old way we created tensor engines via rust local mesh. We were no longer testing this in test_remote_functions, etc. but still have a bunch of old examples that depended on it.

The reason to do this now is to lighten what we actually use in the monarch_tensor_worker crate. We need to sever its direct dependence on torch and getting rid of its dead code makes it clearer what we have to do.

Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1947
ghstack-source-id: 324814292

Differential Revision: [D87496994](https://our.internmc.facebook.com/intern/diff/D87496994/)