Stop the proc mesh from the alloc in HostMesh::shutdown
Summary:
In order to stop the remote process allocation, we need to call `Alloc::stop`, which is
done when the ProcMesh created from the allocation is itself stopped.
This should ensure jobs finish entirely.

Note that the Drop for ProcMesh and Drop for RemoteProcessAlloc both do not try to stop
the underlying allocation, which is why this is necessary.

Reviewed By: pzhan9

Differential Revision: D87264086