[struct_diff_patch] diffing / patching for rust structs
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1989

The crate struct_diff_patch provides "diff" and "patch" operations for rust data types. It defines two generic traits: 1) `Diff` which implements diffing two values to produce a patch; and 2) `Patch` which is implemented on patches in order to apply them to values.

Additionally, the crate provides a "watch", which lets subscribers track changes to a value as a value snapshot + stream of patches.

The intention is to use this in the monarch resource model to observe resources efficiently, without having to implement special operations for each access pattern.

In the future, we can also implement patch _merge_ operations in order to allow patches to be used in accumulations as well.
ghstack-source-id: 325529472
@exported-using-ghexport

Differential Revision: [D87822791](https://our.internmc.facebook.com/intern/diff/D87822791/)