Add torch-requirements file for the nightly version
Summary:
Fixes https://github.com/meta-pytorch/monarch/issues/1978

Separate out the torch requirement into a different file, since it requires a different
index. We don't want `build-requirements.txt` to overwrite it.

It is not used in CI because we use "matrix.torch-spec" to change the torch version we
download. We should ensure the requirements file always matches the tested version.

Differential Revision: D87802441