include user mesh name and endpoint name in supervision error (#1917)
Summary:

The current error message logs the actor name and stacktrace when an exception happens (ActorError/SupervisionError). The endpoint name is included in the stacktrace.

But for cases like proc crash, the supervision error doesn't have a stacktrace, so it couldn't show the endpoint name.

This diff includes the user actor name and endpoint name to the exceptions ActorError and SupervisionError.


github issue: https://github.com/meta-pytorch/monarch/issues/1899

Reviewed By: zdevito

Differential Revision: D87353113