Quick fix for T246995730
Summary: Calling `hyperactor_telemetry::env::execution_url()` in unit tests is causing hangs. A line-by-line bisection shows that its this downstream graphql call that causes this https://fburl.com/code/b99kgut6. The simplest fix is to just not print it for now so we don't need to fully back out of D87273699 and make it easy to forward fix the full functionality

Reviewed By: dulinriley

Differential Revision: D88190271