[hyperactor] ensure that the host agent is always ready when serving a host; remove execution URL
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2030

Ensure that a host agent is always ready when serving a host. Currently, we serve the channel -- making the address available -- and *then* launch the agent. Thus, there is a small race condition where we could receive messages on the channel without the agent being available.

Starting Meta binaries is already slow, and the execution URL adds to it. We also remove this to speed startup.
ghstack-source-id: 326753602
@exported-using-ghexport

Differential Revision: [D88208166](https://our.internmc.facebook.com/intern/diff/D88208166/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D88208166/)!