Update base for Update on "[hyperactor] clarify parsing of names and identifiers; make `Name` concrete syntax unambiguous"
A previous change, D83495465, changed the `Name` format to separate identifiers and uuids with a '_' character. This introduces a parsing ambiguity, because we allow '_' to appear in names, and thus we can no longer unambiguously distinguish between a `Reserved` and `Suffixed` name.  We ran into this in practice because we had actors that used '_' characters in their names.

This change makes parsing lawful: to make the `Name` representation unambiguous, we reintroduce the '-' separator, and permit it as a name *element*. We retain the previous definition of an *identifier*, and apply this check to names that are constructed via `Name::new` and `Name::new_reserved`, making these functions fallible.

Differential Revision: [D88277771](https://our.internmc.facebook.com/intern/diff/D88277771/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D88277771/)!

[ghstack-poisoned]