Fix deadlock in intern_typename
Summary:
There can be a deadlock in DashMap::entry if it is called while any other references are alive.
To avoid this, separate the query for live keys from the insertion path, so the reference is
not alive. There's a small performance cost on the first insert, but the read path still only
has a single query.
Should be negligible, and helps avoid deadlocks.

Differential Revision: D88891169