[monarch] build: detect empty parent
`-print-file-name` returns the path if it is not resolved in the compiler's search path:

```
$ c++ -print-file-name=libstdc++.a
libstdc++.a
```

which means that if the library is missing, we'll compute an empty search path, and given the mysterious error:

```
error: empty search path given via `-L`
```

we now detect this case, and omit the search path altogether, improving the error somewhat:

```
error: could not find native static library `stdc++`, perhaps an -L flag is missing?
```

Possibly we should have the build script actually error out on this case.

Differential Revision: [D88964670](https://our.internmc.facebook.com/intern/diff/D88964670/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D88964670/)!

ghstack-source-id: 328741510
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2125