Adding working dockerfile with overrideable pytorch base image. (#2131)
Summary:

This diff adds a working Dockerfile.  It has been tested w/ the TorchX kubernetes integration on AWS multi-node.

It has an overrideable base image for the nightly from pytorch.

The below test plan shows how to log into a devgpu on the cloud, build it, upload it to the ECR registry, test it, then build with an override and test again.

A good next step is to automate this with GitHub actions.  I feel that building it based on the nightlies is a good move and we should not build this every commit / we should not redo all the build logic separately here as it would be complex and redundant.

Reviewed By: thedavekwon

Differential Revision: D89013574