Modifying commands.py to avoid race condition on scheduling with intermediate states like UNKNOWN in kubernetes.
Summary:
The existing commands.py code for the server_ready function has a failure scenario the way it is currently coded.  This manifests itself roughly 1 in 3 times with the kubernetes scheduler when it is used without modifications.

The issue is that the current code only handles these states:

```
UNSUBMITTED = 0
SUBMITTED = 1
PENDING = 2
RUNNING = 3
```

In kubernetes, maybe 1 in 3 tries, we get this transient state which provides little information, but is not a problem.

```
UNKNOWN = 7
```

There is also a possibility that we get any of these terminal states depending on what ends up happening in launched code:

```
SUCCEEDED = 4
FAILED = 5
CANCELLED = 6
```

This change checks for terminal states and returns them properly, and waits until the next cycle for transient states.

Differential Revision: D89016333