Add enum BindSpec and use it in enable_transport (#2140)
Summary:


As explained in D89190085, Lightning wants to explicitly set the frontend address both client and worker hosts. Since the client's frontend address is controlled through `enable_transport`, this diff adds a new enum BindSpec, and plumb it to `enable_transport`, so the user can pass an explicit address string to do that.

Reviewed By: mariusae

Differential Revision: D89190087