[monarch] make client this_proc() real
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/2137

This builds on the previous change to implement a *real* in-process host and local proc mesh. Thus, nothing is fake anymore in the v1 path: this_proc() behaves the same everywhere, and is managed in precisely the same way.

This also means that we can do things like hand out references to the local host (or proc), and have remove actors spawn (procs, actors) on it!

It will also simplify integration: since the client proc is now a host, it has a single front-end address that multiplexes all its procs (including the local one).

We keep around the old codepaths for the benefit of v0; this can be significantly simplified again once we drop v0 support.
ghstack-source-id: 329960116

Differential Revision: [D89089836](https://our.internmc.facebook.com/intern/diff/D89089836/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D89089836/)!