Move supervision monitor to controller actor (#2157)
Summary:

Part of: https://github.com/meta-pytorch/monarch/issues/2027

Before this change, the actor_states_monitor tokio task was only running on PyActorMesh, and
only monitored PythonActors. Plus it only forwarded the SupervisionFailureMessage to other
PythonActors. This means that rust-based actors have none of this error propagation working.

This change moves the monitor task into the ActorMeshController, which will allow every actor,
regardless of type, to monitor the mesh it owns for failures. Each Ref will subscribe to messages
from the controller to alert it to changes in state. This will also scale better, as instead of each
Mesh and Ref object having a monitor, there is only one per mesh globally. So there should be
fewer messages going around. The tokio task became an actor self-message, which has
better tracking of failures and guarantees, plus no need for async locks on health state.
This also gives more implementation flexibility because the details are inside the controller,
and users only need the `Subscribe` message.
The subscribers are guaranteed to also get a None message, which means that the owner
is still alive, and the actors have not changed their state. This can be used to detect cases
where the controller is unreachable.

To do the forwarding, we require the spawning context of ProcMesh::spawn to
have an impl Handler<SupervisionFailureMessage>.
This message represents when an actor on a mesh your actor (or client) owns fails.
PythonActor uses this message to call its `__supervise__` callback, and rust actors have
no default implementation. The TestRootClient and GlobalRootClient both get implementations
that just panic. In the future we may want something like `unhandled_fault_hook` for Rust
actors.

The new method is `next_supervision_event`, a future that resolves if an event occurs. This
can be awaited simultaneously via tokio::select with a reply from a casted message to re-create
the Python behavior.

Differential Revision: D87491033