Refactor actor init message
This makes it more clear that the actor init message can be creating at the same time as the call to allocate the python actor by refactoring ActorMesh._create. It gets inlined into the ProcMesh spawn code, and calls a method that returns the PythonMessage that is the init message.

However, we currently _cannot_ remove the actual cast of the init message. This is because actor init itself uses the context of the init message to initialize its rank. This information is not available to the rust actor's init method so we cannot handle our init method from within it.

Nevertheless, this change makes it clearer how to merge the init when it is possible.

Differential Revision: [D89700862](https://our.internmc.facebook.com/intern/diff/D89700862/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D89700862/)!

[ghstack-poisoned]